Remove invalid test vectors (from sp1). In the MulChip's eval function, the constraint that checks if the calculated product matches the witnessed result a is guarded by the op_a_not_0 flag.
However, looking at the generate_trace and event_to_row functions, the op_a_not_0 column is never set. It defaults to zero, which effectively disables the constraint. The prover, therefore, has no way of detecting that the value of a is incorrect.

i enabled this constraint by setting op_a_not_0 to 1 in the trace generation.